repair ORM tests that use deferrable FKs
after the bound metadata merge the unitofwork.TransactionTest
became deprecations.AutocommitClosesOnFail and was found to
be not actually running correctly on SQLite.  This test
should only run on a backend that has deferrable FKs,
which includes PostgreSQL; so add that to requirements.  this
in turn requires clarification of test_naturalpks that the
"deferrable FKs" req there is more about Oracle and its lack
of ON UPDATE cascades.

Change-Id: I867a164679274940803852c1af0920da74683bfa